requirements.txt
requests>=2.28.2
uvicorn>=0.20.0
fastapi>=0.89.0
pyYAML>=6.0.1
pandas>=2.1.3